Structre Update

+++ src/main/java/com/rupesh/ticket_management/repository/TicketRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rupesh.ticket_management.entity.Ticket;
 
-public interface TicketRepo extends JpaRepository<Ticket, Integer> {
+public interface TicketRepo extends JpaRepository<Ticket, Long> {
 
 }
