+++ src/main/java/com/rupesh/ticket_management/service/UserService.java
@@ -9,7 +9,7 @@ import com.rupesh.ticket_management.entityDto.response.UserResponseDTO;
 public interface UserService {
 	
 	ResponseEntity<String> addUser(UserDTO user);
-	ResponseEntity<UserResponseDTO> getUser(int Id);
+	ResponseEntity<UserResponseDTO> getUser(Long Id);
